feat(#236): 스터디 검색 API 카테고리 필터 구현
- API, DataSource, Repository, UseCase 구현
- 타입 Int -> Long 변경

+++ app/src/main/java/com/ssafy/neegongnaegong/data/datasource/network/NetworkStudiesDataSource.kt
@@ -103,7 +103,7 @@ interface NetworkStudiesDataSource {
     fun getSearchStudies(
         keyword: String,
         sort: String?,
-        categoryIds: List<Int>?,
+        categoryIds: List<Long>?,
         cursorValue: String?,
         cursorId: Long?,
         size: Int,
