+++ app/src/main/java/com/ssafy/neegongnaegong/data/datasource/network/NetworkStudiesDataSourceImpl.kt
@@ -209,7 +209,7 @@ class NetworkStudiesDataSourceImpl
         override fun getSearchStudies(
             keyword: String,
             sort: String?,
-            categoryIds: List<Int>?,
+            categoryIds: List<Long>?,
             cursorValue: String?,
             cursorId: Long?,
             size: Int,
