+++ app/src/main/java/com/ssafy/neegongnaegong/data/repository/StudiesRepositoryImpl.kt
@@ -256,7 +256,7 @@ class StudiesRepositoryImpl
         override fun getStudiesSearch(
             keyword: String,
             sort: String?,
-            categoryIds: List<Int>?,
+            categoryIds: List<Long>?,
             cursorValue: String?,
             cursorId: Long?,
             size: Int,
